# tree: use FastIntSet during typechecking