@chengxiong-ruan 2022-10-10
Nice work! I tried that as well but didn't see much improvement on cpu usage for function expressions through the two benchmarks I was interested at so I turned to other directions to improve function expressions. Then I added benchmark for operators as well and I forgot to run the operator benchmarks on that change :( Thanks for working on this. This pr show great improvement on allocation for operators!!!

```
BENCHTIMEOUT=24h PKG=./pkg/bench BENCHES=BenchmarkFuncExprTypeCheck ./scripts/bench ffab421f7d623065d768e3be4c3af18f444d3d3e a9d3c81c51ad86528bcc6ded94ebd20de7f0419d

name                                                       old time/op    new time/op    delta
FuncExprTypeCheck/builtin_called_on_null_input-24            1.05µs ± 0%    1.05µs ± 0%   -0.42%  (p=0.001 n=10+10)
FuncExprTypeCheck/builtin_not_called_on_null_input-24        1.00µs ± 1%    1.00µs ± 1%   -0.79%  (p=0.001 n=10+9)
FuncExprTypeCheck/builtin_aggregate-24                       2.16µs ± 1%    2.12µs ± 1%   -2.23%  (p=0.000 n=10+10)
FuncExprTypeCheck/builtin_aggregate_not_called_on_null-24    1.23µs ± 0%    1.20µs ± 1%   -2.44%  (p=0.000 n=10+9)
FuncExprTypeCheck/udf_same_name_as_builtin-24                1.50µs ± 0%    1.53µs ± 0%   +1.72%  (p=0.000 n=10+9)
FuncExprTypeCheck/udf_across_different_schemas-24            2.43µs ± 1%    2.50µs ± 1%   +2.49%  (p=0.000 n=8+10)
FuncExprTypeCheck/unary_operator-24                           345ns ± 2%     345ns ± 2%     ~     (p=1.000 n=10+10)
FuncExprTypeCheck/binary_operator-24                         1.35µs ± 0%    1.31µs ± 1%   -3.22%  (p=0.000 n=10+10)
FuncExprTypeCheck/comparison_operator-24                     2.17µs ± 0%    2.16µs ± 0%   -0.73%  (p=0.000 n=10+10)
FuncExprTypeCheck/tuple_comparison_operator-24               6.49µs ± 0%    6.39µs ± 1%   -1.46%  (p=0.000 n=10+10)
FuncExprTypeCheck/tuple_in_operator-24                       1.13µs ± 1%    1.04µs ± 1%   -7.88%  (p=0.000 n=10+10)

name                                                       old alloc/op   new alloc/op   delta
FuncExprTypeCheck/builtin_called_on_null_input-24              132B ± 7%      117B ± 8%  -11.16%  (p=0.000 n=10+9)
FuncExprTypeCheck/builtin_not_called_on_null_input-24          129B ± 1%      115B ± 0%  -10.94%  (p=0.000 n=8+8)
FuncExprTypeCheck/builtin_aggregate-24                         467B ± 0%      443B ± 0%   -5.16%  (p=0.000 n=9+9)
FuncExprTypeCheck/builtin_aggregate_not_called_on_null-24      145B ± 0%      131B ± 1%   -9.69%  (p=0.000 n=9+8)
FuncExprTypeCheck/udf_same_name_as_builtin-24                  273B ± 0%      261B ± 1%   -4.60%  (p=0.000 n=9+9)
FuncExprTypeCheck/udf_across_different_schemas-24              674B ± 0%      662B ± 0%   -1.78%  (p=0.000 n=8+7)
FuncExprTypeCheck/unary_operator-24                           0.00B          0.00B          ~     (all equal)
FuncExprTypeCheck/binary_operator-24                           113B ± 0%       89B ± 1%  -21.63%  (p=0.000 n=7+9)
FuncExprTypeCheck/comparison_operator-24                       153B ± 1%      129B ± 1%  -15.59%  (p=0.000 n=9+9)
FuncExprTypeCheck/tuple_comparison_operator-24                 893B ± 0%      821B ± 0%   -8.00%  (p=0.000 n=9+9)
FuncExprTypeCheck/tuple_in_operator-24                         400B ± 0%      344B ± 0%  -13.97%  (p=0.000 n=8+9)

name                                                       old allocs/op  new allocs/op  delta
FuncExprTypeCheck/builtin_called_on_null_input-24              7.00 ± 0%      6.00 ± 0%  -14.29%  (p=0.000 n=10+10)
FuncExprTypeCheck/builtin_not_called_on_null_input-24          7.00 ± 0%      6.00 ± 0%  -14.29%  (p=0.000 n=10+10)
FuncExprTypeCheck/builtin_aggregate-24                         16.0 ± 0%      14.0 ± 0%  -12.50%  (p=0.000 n=10+10)
FuncExprTypeCheck/builtin_aggregate_not_called_on_null-24      8.00 ± 0%      7.00 ± 0%  -12.50%  (p=0.000 n=10+10)
FuncExprTypeCheck/udf_same_name_as_builtin-24                  11.0 ± 0%      10.0 ± 0%   -9.09%  (p=0.000 n=10+9)
FuncExprTypeCheck/udf_across_different_schemas-24              16.0 ± 0%      15.0 ± 0%   -6.25%  (p=0.002 n=8+10)
FuncExprTypeCheck/unary_operator-24                            0.00           0.00          ~     (all equal)
FuncExprTypeCheck/binary_operator-24                           5.00 ± 0%      3.00 ± 0%  -40.00%  (p=0.000 n=10+10)
FuncExprTypeCheck/comparison_operator-24                       5.00 ± 0%      3.00 ± 0%  -40.00%  (p=0.000 n=10+10)
FuncExprTypeCheck/tuple_comparison_operator-24                 19.0 ± 0%      13.0 ± 0%  -31.58%  (p=0.000 n=10+10)
FuncExprTypeCheck/tuple_in_operator-24                         7.00 ± 0%      4.00 ± 0%  -42.86%  (p=0.000 n=10+10)
```

```
BENCHTIMEOUT=24h PKG=./pkg/bench BENCHES=BenchmarkSQL/MultinodeCockroach/Upsert/count=1000 ./scripts/bench ffab421f7d623065d768e3be4c3af18f444d3d3e a9d3c81c51ad86528bcc6ded94ebd20de7f0419d

name                                         old time/op    new time/op    delta
SQL/MultinodeCockroach/Upsert/count=1000-24    23.7ms ± 3%    23.7ms ± 4%   ~     (p=0.912 n=10+10)

name                                         old alloc/op   new alloc/op   delta
SQL/MultinodeCockroach/Upsert/count=1000-24    10.1MB ±11%    10.2MB ±18%   ~     (p=0.842 n=9+10)

name                                         old allocs/op  new allocs/op  delta
SQL/MultinodeCockroach/Upsert/count=1000-24     53.1k ±13%     51.1k ±11%   ~     (p=0.190 n=10+10)
```