@jordanlewis 2022-10-10
Awesome, thanks for running your benchmarks again and good to know! I'll add a backport tag, but not sure if that's desirable - let me know.

bors r+